210630 | MVVM, AAC로 코드 리팩토링|ROOM + LiveData + ViewModel + DataBinding

+++ app/src/main/java/com/heon9u/alarm_weather_app/alarm/AlarmListView.java
@@ -1,7 +1,10 @@
 package com.heon9u.alarm_weather_app.alarm;
 
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
+import android.graphics.Color;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -38,8 +41,10 @@ import com.heon9u.alarm_weather_app.databinding.AlarmFragmentBinding;
 import java.util.List;
 
 public class AlarmListView extends Fragment {
+    public static final int CREATE_ALARM_REQUEST = 1;
+    public static final int UPDATE_ALARM_REQUEST = 2;
 
-    private AlarmViewModel alarmViewModel;
+    public AlarmViewModel alarmViewModel;
     public AlarmFragmentBinding alarmFragmentBinding;
     Context context;
     RecyclerView recyclerView;
@@ -59,7 +64,6 @@ public class AlarmListView extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
-        // RecyclerView로 alarm 리스트 페이지 호출
         alarmFragmentBinding = DataBindingUtil.inflate(inflater, R.layout.alarm_fragment, container, false);
         alarmFragmentBinding.setAlarmClick(this);
         View view = alarmFragmentBinding.getRoot();
@@ -84,6 +88,24 @@ public class AlarmListView extends Fragment {
             }
         });
 
+        alarmAdapter.setOnItemClickListener(alarm -> {
+            Intent updateIntent = new Intent(getActivity(), AlarmSetActivity.class);
+            updateIntent.putExtra("alarm", alarm);
+            updateIntent.putExtra("REQUEST_STATE", "update");
+            startActivityForResult(updateIntent, UPDATE_ALARM_REQUEST);
+        });
+
+        alarmAdapter.setOnCheckedChangeListener((alarm, isChecked) -> {
+            alarm.setTotalFlag(isChecked);
+            alarmViewModel.update(alarm);
+
+            if(isChecked) {
+                changeAlarmOnOff(alarm, "reboot");
+            } else {
+                changeAlarmOnOff(alarm, "cancel");
+            }
+        });
+
         return view;
     }
 
@@ -95,12 +117,26 @@ public class AlarmListView extends Fragment {
     public void createAlarmClick(View view) {
         Intent createAlarmIntent = new Intent(getActivity(), AlarmSetActivity.class);
         createAlarmIntent.putExtra("REQUEST_STATE", "create");
-        startActivity(createAlarmIntent);
+        startActivityForResult(createAlarmIntent, CREATE_ALARM_REQUEST);
     }
 
     public void manageLocationClick(View view) {
         Intent menuLocationIntent = new Intent(getActivity(), LocationListView.class);
         startActivity(menuLocationIntent);
+    }
+
+    public void registeredAlarmManager(String request, Alarm alarm) {
+        Intent alarmIntent = new Intent(getActivity(), AlarmManagerActivity.class);
+        alarmIntent.putExtra("alarm", alarm);
+        alarmIntent.putExtra("request", request);
+        startActivity(alarmIntent);
+    }
+
+    public void changeAlarmOnOff(Alarm alarm, String request) {
+        Intent alarmIntent = new Intent(context, AlarmManagerActivity.class);
+        alarmIntent.putExtra("alarm", alarm);
+        alarmIntent.putExtra("request", request);
+        context.startActivity(alarmIntent);
     }
 
     public void attachItemTouchHelperToAdapter() {
@@ -113,10 +149,56 @@ public class AlarmListView extends Fragment {
 
             @Override
             public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-                alarmViewModel.delete(alarmAdapter.getAlarmAt(viewHolder.getBindingAdapterPosition()));
-                Toast.makeText(getContext(), "알람을 삭제했습니다.", Toast.LENGTH_SHORT).show();
+                showDialogDeleteAlarm(viewHolder);
             }
+
         }).attachToRecyclerView(recyclerView);
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if(resultCode != -1) return;
+
+        Alarm alarm = (Alarm) data.getSerializableExtra("alarm");
+        Log.e("Alarm", alarm.toString());
+
+        switch(requestCode) {
+            case CREATE_ALARM_REQUEST:
+                alarmViewModel.insert(alarm);
+                return;
+            case UPDATE_ALARM_REQUEST:
+                alarmViewModel.update(alarm);
+                return;
+        }
+
+        registeredAlarmManager("create", alarm);
+    }
+
+    public void showDialogDeleteAlarm(@NonNull RecyclerView.ViewHolder viewHolder) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+        builder.setTitle("알람 삭제")
+                .setMessage("해당 알람을 삭제하시겠습니까?")
+                .setIcon(android.R.drawable.ic_menu_delete)
+                .setCancelable(false)
+                .setPositiveButton("삭제", (dialog, which) -> {
+                    alarmViewModel.delete(alarmAdapter.getAlarmAt(viewHolder.getBindingAdapterPosition()));
+                    Toast.makeText(getContext(), "알람을 삭제했습니다.", Toast.LENGTH_SHORT).show();
+                })
+                .setNegativeButton("취소", (dialog, which) -> {
+                    alarmAdapter.notifyItemChanged(viewHolder.getBindingAdapterPosition());
+                    Toast.makeText(getContext(), "삭제를 취소했습니다.", Toast.LENGTH_SHORT).show();
+                }).show();
+    }
+
+    @Override
+    public void onDestroy() {
+        if(nativeAd != null) {
+            nativeAd.destroy();
+        }
+
+        super.onDestroy();
     }
 
     public void initAdMob() {
@@ -128,7 +210,7 @@ public class AlarmListView extends Fragment {
                 nativeAd = unifiedNativeAd;
 
             LayoutInflater inflater = (LayoutInflater) context
-                                            .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+                    .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
             UnifiedNativeAdView adView = (UnifiedNativeAdView) inflater
                     .inflate(R.layout.native_ad_layout, null);
             populateNativeAd(unifiedNativeAd, adView);
@@ -158,14 +240,5 @@ public class AlarmListView extends Fragment {
         }
 
         adView.setNativeAd(nativeAd);
-    }
-
-    @Override
-    public void onDestroy() {
-        if(nativeAd != null) {
-            nativeAd.destroy();
-        }
-
-        super.onDestroy();
     }
 }
